\name{dCovTS-package}
\alias{dCovTS-package}
\alias{dCovTS}
\docType{package}
\title{
Distance Covariance and Correlation theory in time series
}
\description{
Computing and plotting the distance covariance and correlation function of a univariate or a multivariate time series. 
Both versions of biased and unbiased estimators of distance covariance and correlation are provided.
Test statistics for testing pairwise independence are also implemented. Some data sets are also included.}
\details{
\tabular{ll}{
Package: \tab dCovTS\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-09-21\cr
License: \tab GPL(>=2)\cr
}
}
\author{
Maria Pitsillou and Konstantinos Fokianos
}
\references{
Dehling, H. and T. Mikosch (1994). Random quadratic forms and the bootstrap for U-statistics. 
\emph{Journal of Multivariate Analysis} \eqn{\textbf{51}}{%
51}, 392-413,
\url{http://dx.doi.org/10.1006/jmva.1994.1069}

Fokianos K. and M. Pitsillou (2016a). Consistent testing for pairwise dependence in time series. \emph{Technometrics},
\url{http://dx.doi.org/10.1080/00401706.2016.1156024}.

Fokianos K. and M. Pitsillou (2016b). Testing pairwise independence for multivariate
time series by the auto-distance correlation matrix. Submitted for publication.

Hong, Y. (1996). Consistent testing for serial correlation of unknown form. \emph{Econometrica} \eqn{\textbf{64}}{%
64}, 837-864,
\url{http://dx.doi.org/10.2307/2171847}.

Hong, Y. (1999). Hypothesis testing in time series via the empirical characteristic function: A generalized
spectral density approach. \emph{Journal of the American Statistical Association} \eqn{\textbf{94}}{%
94}, 1201-1220,
\url{http://dx.doi.org/10.1080/01621459.1999.10473874}.

Leucht, A. and M. H. Neumann (2013). Dependent wild bootstrap for degenerate U- and V- statistics.
\emph{Journal of Multivariate Analysis} \eqn{\textbf{117}}{%
117}, 257-280,
\url{http://dx.doi.org/10.1016/j.jmva.2013.03.003}.

Politis, N. P., J. P. Romano and M. Wolf (1999). \emph{Subsampling}. New York: Springer.

Shao, X. (2010). The dependent wild bootstrap.
\emph{Journal of the American Statistical Association} \eqn{\textbf{105}}{%
105}, 218-235,
\url{http://dx.doi.org/10.1198/jasa.2009.tm08744}.

Shumway, R. H. and D. S. Stoffer (2011). \emph{Time Series Analysis and Its Applications With R Examples}. New York: Springer.
Third Edition. \url{http://www.stat.pitt.edu/stoffer/tsa3/}

Szekely, G. J. and M. L. Rizzo (2014). Partial distance correlation with methods for dissimilarities. 
\emph{The Annals of Statistics} \eqn{\textbf{42}}{%
42}, 2382-2412,
\url{dx.doi.org/10.1214/14-AOS1255}.

Szekely, G. J., M. L. Rizzo and N. K. Bakirov (2007). Measuring and testing dependence
by correlation of distances. \emph{The Annals of Statistics} \eqn{\textbf{35}}{%
35}, 2769-2794,
\url{http://dx.doi.org/10.1214/009053607000000505}.

Tsay, R. S. (2010). \emph{Analysis of Financial Time Series}. Hoboken, NJ: Wiley. Third edition.

Tsay, R. S. (2014). \emph{Multivariate Time Series Analysis with R and Financial Applications}. Hoboken, NJ:
Wiley.

Zhou, Z. (2012). Measuring nonlinear dependence in time series, a distance correlation approach.
\emph{Journal of Time Series Analysis} \eqn{\textbf{33}}{%
33}, 438-457,
\url{http://dx.doi.org/10.1111/j.1467-9892.2011.00780.x}.
}
\keyword{ package }


